+++ package.json
@@ -7,10 +7,10 @@
   "scripts": {
     "lint": "eslint",
     "lint:fix": "eslint --fix",
-    "unit": "node --test test.js",
+    "unit": "node --test test/*.js",
     "test:typescript": "tsd",
     "test": "npm run lint && npm run unit && tsd",
-    "coverage": "c8 --reporter=lcov node --test test.js",
+    "coverage": "c8 --reporter=lcov npm run unit",
     "coverage:report": "c8 report",
     "test:ci": "npm run lint && npm run coverage && npm run test:typescript",
     "license-checker": "license-checker --production --onlyAllow='MIT;ISC;BSD-3-Clause;BSD-2-Clause'",
